master-ip/image_search/requirements.txt
torch
transformers
pillow
pymongo
pinecone
tqdm